Optimising SendMail.js
