+++ .env.example
@@ -1,5 +1,5 @@
 PORT=3000
 MONGO_URI=MongoDB-URI
 JWT_SECRET=YOUR_JWT_SECRET
-EMAIL_USER=YOUR_EMAIL_ADDRESS
-EMAIL_PASS=YOUR_APP_PASSWORD
+SENDER_EMAIL=YOUR_EMAIL_ADDRESS
+SENDER_APP_PASSWORD=YOUR_APP_PASSWORD
